+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Highlights are marked with a pancake 🥞
 - Fix SQLite store handling of `previous` hashes [#1051](https://github.com/p2panda/p2panda/pull/1051)
 - Fix missing gossip events in sync manager [#988](https://github.com/p2panda/p2panda/pull/988)
 
+### Added
+
+- Add `TopicStore` trait and SQLite implementation [#1011](https://github.com/p2panda/p2panda/pull/1011)
+
 ## [0.5.1] - 09/02/2026
 
 ### Changed